ref(sampling): Remove dependency from sampling to event schema
